requirements.txt
django>=2.0,<4.0
django-filter
django-nested-admin>=3.3.2
django-taggit
django-taggit-serializer
pillow>=7.2.0
python-magic>=0.4.18
pymongo>=3.11.0
django-rest-framework
uritemplate
pyyaml
pydantic
django-htmlmin>=0.11.0
django-redis>=4.12.1
unicms-template-italia
markdown
markdownify
